Fixes and Explosions
Added the ability to add an optional explosion to when the bullet hits
an object and to when an AI dies. Added the HUD Manager script to allow
the display of a timer, countdown, health (Option pseudo graphical),
the score and the ammo on current gun. Added to the instructions.